requirements.txt
ccxt>=4.5.4
alpaca-py==0.43.1
redis>=5.2.1
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.19
pydantic>=2.10.4,<3.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
numpy>=2.2.0,<3.0
pandas>=2.2.0,<3.0
tenacity>=8.5.0
aiohttp>=3.10.11
httpx>=0.27.0
uvicorn>=0.30.0
fastapi>=0.111.0
structlog>=24.1.0
prometheus-client>=0.21.0
-e ./vendor/freqtrade
-e ./vendor/freqtrade
